+++ src/CodeNexus.Domain/Entities/LearningPathShare.cs
@@ -9,9 +9,9 @@ namespace CodeNexus.Domain.Entities
         public Guid MentorId { get; set; }
         public Guid StudentId { get; set; }
         public Guid? AcceptedPathId { get; set; }
-        public int? SourceVersionAtAccept { get; set; }
-        public int? IgnoredSourceVersion { get; set; }
-        public int? LastNotifiedSourceVersion { get; set; }
+        public decimal? SourceVersionAtAccept { get; set; }
+        public decimal? IgnoredSourceVersion { get; set; }
+        public decimal? LastNotifiedSourceVersion { get; set; }
         public bool IsTrackingEnabled { get; set; } = true;
         public string? InvalidatedReason { get; set; }
         public LearningPathShareStatus Status { get; set; } = LearningPathShareStatus.Pending;
